double tile object coords

+++ src/main/java/core/resources/tilemap/TileObject.java
@@ -8,8 +8,8 @@ public class TileObject {
     public int id;
     public String name = "";
     public int gid = 0; // Tile id
-    public int x, y;
-    public int width, height;
+    public double x, y;
+    public double width, height;
     public String type = "";
     public double rotation = 0.0;
     public boolean visible = true;
